Intergrated ESI, allowing users to allocate ESI Network quotas
A new option `--esi` has been added to the `add_openstack_resource` command,
allowing the creation of ESI resources in Coldfront.

Allocations of ESI resources will be initialized with only a network quota
of 1 floating IP and network by default.

The implementation of this feature required subclassing of the
`OpenstackResourceAllocator` to create an allocator class for
ESI resources, changes to the various Coldfront commands,
the addition of one Openstack quota attribute,a small
modification to the CI and test files, and the addition of a test
file for ESI allocations.